# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get a list of employees",
                "description": "Get a list of employees.",
                "operationId": "9aeb15c3a8c5b0b8b9e0931d1673e787",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create a new employee",
                "description": "Create a new employee.",
                "operationId": "2475af536398d342aca2fd7f60f9af3a",
                "responses": {
                    "201": {
                        "description": "Employee created successfully"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/employees/{employee}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Show the specified employee",
                "description": "Show the specified employee.",
                "operationId": "c2a3a0ec2a75e78c2d9e622df0085a77",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update the specified employee",
                "description": "Update the specified employee.",
                "operationId": "6e77188ef152be1ee6ec3321fcb02882",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete the specified employee",
                "description": "Delete the specified employee.",
                "operationId": "799db2e2d0fe3a7fb4c8dd8adb3484e7",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}